+++ tests/svrunit/fixtures/custom-command/autoload.php
@@ -1,0 +1,5 @@
+<?php
+
+include_once __DIR__ . '/src/CustomCommand.php';
+
+\PHPUnuhi\AppManager::registerExtensionCommand(new CustomCommand());